fix: use C.size_t instead of C.ulong for Windows LLP64 compatibility
On Windows, unsigned long is 32-bit but size_t is 64-bit. The Rust FFI
signature uses usize (mapped to size_t by cbindgen), so C.size_t is
the correct type on all platforms.

+++ pkg/internal/table.go
@@ -184,7 +184,7 @@ func (t *Table) AddRecords(_ context.Context, records []arrow.Record, _ *contrac
 		t.handle,
 		// #nosec G103 - Safe conversion of Go slice to C array pointer for FFI
 		(*C.uchar)(unsafe.Pointer(&ipcBytes[0])),
-		C.ulong(len(ipcBytes)),
+		C.size_t(len(ipcBytes)),
 		&addedCount,
 	)
 	defer C.simple_lancedb_result_free(result)
